requirements.txt
Django==1.9.12
djangorestframework==3.5.4
kafka-python==1.3.5
MySQL-python==1.2.5
